propagate float and stop using it too much

+++ model-integration/src/main/java/ai/vespa/rankingexpression/importer/onnx/TypeConverter.java
@@ -53,17 +53,17 @@ class TypeConverter {
 
     private static TensorType.Value toValueType(Onnx.TensorProto.DataType dataType) {
         switch (dataType) {
-            case FLOAT: return TensorType.Value.FLOAT;
+            case FLOAT: return TensorType.Value.DOUBLE;
             case DOUBLE: return TensorType.Value.DOUBLE;
             // Imperfect conversion, for now:
-            case BOOL: return TensorType.Value.FLOAT;
-            case INT8: return TensorType.Value.FLOAT;
-            case INT16: return TensorType.Value.FLOAT;
-            case INT32: return TensorType.Value.FLOAT;
+            case BOOL: return TensorType.Value.DOUBLE;
+            case INT8: return TensorType.Value.DOUBLE;
+            case INT16: return TensorType.Value.DOUBLE;
+            case INT32: return TensorType.Value.DOUBLE;
             case INT64: return TensorType.Value.DOUBLE;
-            case UINT8: return TensorType.Value.FLOAT;
-            case UINT16: return TensorType.Value.FLOAT;
-            case UINT32: return TensorType.Value.FLOAT;
+            case UINT8: return TensorType.Value.DOUBLE;
+            case UINT16: return TensorType.Value.DOUBLE;
+            case UINT32: return TensorType.Value.DOUBLE;
             case UINT64: return TensorType.Value.DOUBLE;
             default: throw new IllegalArgumentException("A ONNX tensor with data type " + dataType +
                                                         " cannot be converted to a Vespa tensor type");
